+++ examples/general/SRL/lib/regex/char_shorthand.rb
@@ -28,10 +28,11 @@ module Regex # This module is used as a namespace
       @shortname = valid_shortname(aShortname)
     end
 
-  public
+    protected
+  
     # Conversion method re-definition.
     # Purpose: Return the String representation of the expression.
-    def to_str()
+    def text_repr()
       return "\\#{shortname}"
     end